\name{rain}
\alias{rain}
\title{Tokyo Rainfall}

\description{

The \code{rain} data has 366 real valued binomial responses
representing days on which it rained over the course of two years in
Tokyo.

}

\format{ 

  A list with three components: \code{y}, the repsonse representing
  the number of times it has rained on the ith day of the year in
  Tokyo from 1983 through 1984, \code{X}, a matrix of ones, and
  \code{n}, the number of observations for each day.  \code{n} is two
  for all days except Feb. 29.

}

\details{

  This is the infamous Tokyo rainfall data from Kitagawa (1987).

}

% \source{}

\references{

Genshiro Kitagawa.  Non-Gaussian State-Space Modeling of Non-stationary
Time Series.  Journal of the American Statistical Association (1987).

}

\keyword{datasets}